pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unbiased"
version = "0.0.2"
description = "Comprehensive bias detection and mitigation for datasets, models, and LLM agents"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "NoBias Team", email = "contact@nobias.dev"}
]
keywords = [
    "bias",
    "fairness",
    "machine-learning",
    "ai",
    "llm",
    "audit",
    "ethics",
    "responsible-ai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
agent = [
    "groq>=0.4.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    "aiohttp>=3.9.0",
]
reports = [
    "reportlab>=4.0.0",
    "matplotlib>=3.7.0",
]
embeddings = [
    "sentence-transformers>=2.2.0",
]
langgraph = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-groq>=0.2.0",
]
server = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]
all = [
    "groq>=0.4.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    "aiohttp>=3.9.0",
    "reportlab>=4.0.0",
    "matplotlib>=3.7.0",
    "sentence-transformers>=2.2.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-groq>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]

[project.urls]
Homepage = "https://github.com/nobias/unbiased"
Documentation = "https://github.com/nobias/unbiased/blob/main/docs"
Repository = "https://github.com/nobias/unbiased"
"Bug Tracker" = "https://github.com/nobias/unbiased/issues"

[tool.setuptools]
package-dir = {"" = "library"}

[tool.setuptools.packages.find]
where = ["library"]
include = ["*"]
exclude = ["*.tests", "*.tests.*", "tests.*", "tests", "*.__pycache__"]

[tool.setuptools.package-data]
"unbiased.agent_audit.personas" = ["data/*.json"]
